Complete TaskList of A-MoreOOP
