+++ src/main/java/parsers/Parser.java
@@ -1,0 +1,5 @@
+package parsers;
+
+public class Parser {
+
+}
